sslvvb

+++ src/main/java/org/example/sorting/MergeSorter.java
@@ -7,9 +7,7 @@ import org.example.strategies.IntReader;
 import org.example.strategies.StringReader;
 import org.example.strategies.ReaderStrategy;
 
-//import java.io.File;
 import java.io.IOException;
-import java.io.FileNotFoundException; //
 import java.io.FileWriter;
 import java.util.ArrayList;
 import java.nio.file.Files;
@@ -17,10 +15,10 @@ import java.nio.file.Paths;
 
 public class MergeSorter {
 
-    private SortMode sortMode;
-    private DataType dataType;
-    private String outputFile;
-    private ArrayList<String> inputFiles;
+    private final SortMode sortMode;
+    private final DataType dataType;
+    private final String outputFile;
+    private final ArrayList<String> inputFiles;
 
     public MergeSorter(Parser parser) {
         this.sortMode = parser.getSortMode();
@@ -29,30 +27,37 @@ public class MergeSorter {
         this.inputFiles = parser.getInputFiles();
     }
 
-    public void startSort() throws IOException { // FileWriter exception
-//        if (Files.exists(Paths.get(outputFile))) {
-//            System.out.println("Output file already exists. Specify the name of a file that does not exist.");
-//            return; // change on exception ?
-//        }
+    public void startSort() throws IOException {
+        if (Files.exists(Paths.get(outputFile))) {
+            throw new IOException("Output file already exists. Specify the name of a file that does not exist.");
+        }
+
         ArrayList<ReaderStrategy> readersList = createReaderList();
+
         FileWriter writeFile = new FileWriter(outputFile);
-        while (readersList.size() > 0) {
-            int minValueIndex = 0;
-            for (int i = 1; i < readersList.size(); ++i) {
-                if (readersList.get(minValueIndex).compareValues(readersList.get(i))) {
-                    minValueIndex = i;
-                }
-            }
-            writeFile.write(readersList.get(minValueIndex).getValue() + "\n");
-            if (readersList.get(minValueIndex).shiftValue() == 1) {
+        while (!readersList.isEmpty()) {
+            int minValueIndex = findMinValueIndex(readersList);
+            ReaderStrategy minValueReader = readersList.get(minValueIndex);
+            writeFile.write(minValueReader.getValue() + "\n");
+            if (minValueReader.shiftValue() != 0) {
                 readersList.remove(minValueIndex);
             }
         }
         writeFile.close();
     }
 
+    private int findMinValueIndex(ArrayList<ReaderStrategy> readersList) {
+        int minValueIndex = 0;
+        for (int i = 1; i < readersList.size(); ++i) {
+            if (readersList.get(minValueIndex).compareValues(readersList.get(i))) {
+                minValueIndex = i;
+            }
+        }
+        return minValueIndex;
+    }
+
     private ArrayList<ReaderStrategy> createReaderList() {
-        ArrayList<ReaderStrategy> readersList = new ArrayList<ReaderStrategy>();
+        ArrayList<ReaderStrategy> readersList = new ArrayList<>();
         for (String fileName : this.inputFiles) {
             try {
                 if (this.dataType == DataType.INTEGER) {
@@ -60,13 +65,10 @@ public class MergeSorter {
                 } else {
                     readersList.add(new StringReader(fileName, this.sortMode));
                 }
-            } catch (FileNotFoundException ex) {
-                System.out.println(ex.getMessage()); // in12.txt (No such file or directory)
-            } catch (Exception ex) {
-                System.out.println(ex.getMessage());
+            } catch (IOException e) {
+                System.out.println(e.getMessage());
             }
         }
         return readersList;
     }
-
 }
